requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
binaryornot==0.4.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.2
django-bootstrap5==24.3
django-filter==24.3
django-taggit==6.1.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
fastapi==0.115.5
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
repath==0.9.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
starlette==0.41.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
watchdog==4.0.2
watchfiles==0.24.0
websockets==14.1
